feat: add UI for configuring webhooks
