+++ frontend/src/views/AdminSettingsPage.vue
@@ -40,7 +40,7 @@ import { onMounted, ref } from 'vue'
 import { useUserStore } from '../store/user.js'
 import { useAdminStore } from '../store/admin.js'
 import { useFormsStore } from '../store/forms.js'
-import WebhookEditModal from '@/components/modals/WebhookEditModal.vue'
+import WebhookEditModal from '@/components/modals/WebhookModal.vue'
 import InformationRequirementsSection from '@/components/InformationRequirementsSection.vue'
 import WebhooksSection from '@/components/WebhooksSection.vue'
 import LoadingIndicator from '@/components/LoadingIndicator.vue'
